IMPALA-11908: Parser change for Iceberg metadata querying
This change extends parsing table references with Iceberg metadata
tables. The TableName class has been extended with an extra vTbl field
which is filled when a virtual table reference is suspected. This
additional field helps to keep the real table in the statment table
cache next to the virtual table, which should be loaded so Iceberg
metadata tables can be created.

Iceberg provides a rich API to query metadata, these Iceberg API tables
are accessible through the MetadataTableUtils class. Using these table
schemas it is possible to create an Impala table that can be queried
later on.

Querying a metadata table at this point is expected to throw a
NotImplementedExpception.

Testing:
 - Added E2E test to test it for some tables.

Change-Id: I0b5db884b5f3fecbd132fcb2c2cbd6c622ff965b