IMPALA-11950: Planner change for Iceberg metadata querying
This commit extends the planner with IcebergMetadataScanNode, this node
is only part of the query plan. At this point it is mostly a
placeholder, will be filled with more data when writing the backend
part.

To avoid executing the plan there is a hardcoded condition, it is after
the explain part, so the change remains testable with EXPLAIN queries.

Testing:
 - Added planner tests

Change-Id: I3675d7a57ca570bfec306798589b5ef6aa34b5c6