IMPALA-11996: Scanner change for Iceberg metadtata querying
This commit adds a scan node for querying Iceberg metadata tables. The
method resembles to HBase table scans, the scan is executed with the
help of the Iceberg API and through JNI.

There is only one fragment created to query the Iceberg Metadata table
which is supposed to be executed on the coordinator node that already
has the metadata ready.

This change will not cover struct column types these slots are se to
NULL, it will be done in IMPALA-12205.

Testing:
 - Added e2e tests for querying metadata tables
 - Updated planner tests

Change-Id: I0e943cecd77f5ef7af7cd07e2b596f2c5b4331e7