IMPALA-12661: Fix ASAN heap-use-after-free in IcebergMetadataScanNode
Thic commit fixes a heap usages after a free that was detected by our
ASAN builds.

During tree traversal the current_type s variablei is reassigned to its
children which is part of the object. However, by the end of the
assignment the rhs object will be destroyed. To fix this issue, the
variable usage was replaced with a pointer.

Testing:
 - Ran tests on ASAN build

Change-Id: I6df9c9cb6914a0c6c93b61aa0dd02acfdba68851